# Add mixed index support for has(key) query