--- a/janusgraph-core/src/main/java/org/janusgraph/graphdb/query/index/AbstractIndexSelectionStrategy.java
+++ b/janusgraph-core/src/main/java/org/janusgraph/graphdb/query/index/AbstractIndexSelectionStrategy.java
@@ -23,6 +23,7 @@ import org.janusgraph.core.schema.SchemaStatus;
 import org.janusgraph.diskstorage.configuration.Configuration;
 import org.janusgraph.graphdb.database.IndexSerializer;
 import org.janusgraph.graphdb.internal.OrderList;
+import org.janusgraph.graphdb.query.JanusGraphPredicate;
 import org.janusgraph.graphdb.query.QueryUtil;
 import org.janusgraph.graphdb.query.condition.*;
 import org.janusgraph.graphdb.query.graph.JointIndexQuery;
@@ -209,7 +210,7 @@ public abstract class AbstractIndexSelectionStrategy implements IndexSelectionSt
             return false;
         }
         final PredicateCondition<RelationType, JanusGraphElement> atom = (PredicateCondition) condition;
-        if (atom.getValue() == null) {
+        if (atom.getValue() == null && atom.getPredicate() != Cmp.NOT_EQUAL) {
             return false;
         }
 
@@ -220,7 +221,11 @@ public abstract class AbstractIndexSelectionStrategy implements IndexSelectionSt
             .filter(field -> field.getStatus() == SchemaStatus.ENABLED)
             .filter(field -> field.getFieldKey().equals(key))
             .findAny().orElse(null);
-        return match != null && indexInfo.supports(index, match, atom.getPredicate());
+        if (match == null) {
+            return false;
+        }
+        boolean existsQuery = atom.getValue() == null && atom.getPredicate() == Cmp.NOT_EQUAL && indexInfo.supportsExistsQuery(index, match);
+        return existsQuery || indexInfo.supports(index, match, atom.getPredicate());
     }
 
     private Map.Entry<Condition,Collection<Object>> getEqualityConditionValues(