--- a/janusgraph-lucene/src/test/java/org/janusgraph/diskstorage/lucene/BerkeleyLuceneTest.java
+++ b/janusgraph-lucene/src/test/java/org/janusgraph/diskstorage/lucene/BerkeleyLuceneTest.java
@@ -51,6 +51,11 @@ public class BerkeleyLuceneTest extends JanusGraphIndexTest {
     }
 
     @Override
+    public boolean supportsGeoPointExistsQuery() {
+        return false;
+    }
+
+    @Override
     public boolean supportsWildcardQuery() {
         return false;
     }