--- a/janusgraph-solr/src/test/java/org/janusgraph/diskstorage/solr/SolrJanusGraphIndexTest.java
+++ b/janusgraph-solr/src/test/java/org/janusgraph/diskstorage/solr/SolrJanusGraphIndexTest.java
@@ -38,6 +38,11 @@ public abstract class SolrJanusGraphIndexTest extends JanusGraphIndexTest {
         return true;
     }
 
+    @Override
+    public boolean supportsGeoPointExistsQuery() {
+        return false;
+    }
+
     @Test
     public void testRawQueries() {
         clopen(option(SolrIndex.DYNAMIC_FIELDS,JanusGraphIndexTest.INDEX),false);