@li-boxuan 2020-08-17
Using the following code snippet and Cassandra + Elasticsearch to compare the performance:
```groovy
graph = JanusGraphFactory.open("conf/janusgraph-cql-es.properties")
mgmt = graph.openManagement()
nameKey = mgmt.makePropertyKey('name').dataType(String.class).make()
mgmt.buildIndex("name", Vertex.class).addKey(nameKey).buildMixedIndex("search")
mgmt.commit()

g = graph.traversal()
for (i = 0; i < 1000; i++) {
g.addV().property("name", "value" + i).iterate();
}
for (i = 0; i < 100000; i++) {
g.addV().iterate();
}
graph.tx().commit()

g.V().has("name").count().profile()
```


With this PR it takes 58 ms.
```
==>Traversal Metrics
Step                                                               Count  Traversers       Time (ms)    % Dur
=============================================================================================================
JanusGraphStep([],[name.neq])                                                                 55.663    95.12
    \_condition=(name <> null)
    \_orders=[]
    \_isFitted=true
    \_isOrdered=true
    \_query=[(name <> null)](2000):name
    \_index=name
    \_index_impl=search
  optimization                                                                                 0.036
  optimization                                                                                 1.727
  backend-query                                                        0                      49.374
    \_query=name:[(name <> null)](2000):name
    \_limit=2000
CountGlobalStep                                                        1           1           2.857     4.88
                                            >TOTAL                     -           -          58.521        -
```

Before this PR, the 0.6.0 SNAPSHOT version: (6b6371d6f)
it takes 125 seconds

```
20:29:29 WARN  org.janusgraph.graphdb.transaction.StandardJanusGraphTx  - Query requires iterating over all vertices [()]. For better performance, use indexes
==>Traversal Metrics
Step                                                               Count  Traversers       Time (ms)    % Dur
=============================================================================================================
JanusGraphStep(vertex,[])                                         101000      101000        4417.593     3.53
    \_condition=()
    \_orders=[]
    \_isFitted=false
    \_isOrdered=true
    \_query=[]
  optimization                                                                                 0.047
  optimization                                                                                 0.030
  scan                                                                                         0.000
    \_query=[]
    \_fullscan=true
    \_condition=VERTEX
TraversalFilterStep([JanusGraphPropertiesStep([...                  1000        1000      120557.293    96.46
  JanusGraphPropertiesStep([name],value)                                                  120343.812
    \_condition=type[name]
    \_orders=[]
    \_isFitted=true
    \_isOrdered=true
    \_query=org.janusgraph.diskstorage.keycolumnvalue.SliceQuery@811bd45d
    \_vertices=1
    optimization                                                                               0.578
    optimization                                                                               0.049
    optimization                                                                               0.052
    .....many many more lines of optimization....
    optimization                                                                               0.002
    optimization                                                                               0.001 
CountGlobalStep                                                        1           1           4.142     0.00
                                            >TOTAL                     -           -      124979.029        -
```