@li-boxuan 2020-09-06
Before rebasing, CI was passing. After rebasing, Travis has failed three times, each time with a different reason: https://travis-ci.org/github/JanusGraph/janusgraph/jobs/723105411, https://travis-ci.org/github/JanusGraph/janusgraph/jobs/723810423, and https://travis-ci.org/github/JanusGraph/janusgraph/jobs/724649659

Can any maintainer with Travis permissions retry the failed job? Thanks!