@rngcntr 2020-09-07
If no maintainer responds, you can retry it yourself by `git commit --amend` without any changes and then `git push -f`. Travis will see this as a new version because `git` gives it a new commit hash. But on the downside, _all_ tests will run again, not only the failed ones.