@porunov 2020-09-07
It was restarted yesterday but maybe not all failed jobs were restarted or they failed again (not sure). I restarted those failed jobs